requirements.txt
openpyxl
python-docx
langchain
langchain-openai
langchain-community
pandas
faiss-cpu
